Add token to requests

+++ src/main/java/com/socialhk/social_network/controller/FriendsController.java
@@ -1,10 +1,8 @@
 package com.socialhk.social_network.controller;
 
-import com.socialhk.social_network.model.entity.FriendsEntity;
-import com.socialhk.social_network.model.entity.FriendsId;
-import com.socialhk.social_network.model.entity.PostId;
-import com.socialhk.social_network.model.entity.UserEntity;
+import com.socialhk.social_network.model.entity.*;
 import com.socialhk.social_network.model.repository.FriendRepository;
+import com.socialhk.social_network.model.repository.TokensRepository;
 import com.socialhk.social_network.model.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -22,27 +20,43 @@ public class FriendsController {
     @Autowired
     UserRepository userRepository;
 
-    @GetMapping(path = "/{id}",produces = "application/json")
+    @Autowired
+    TokensRepository tokensRepository;
+
+    @GetMapping(path = "/{id}/{token}",produces = "application/json")
     @ResponseStatus(code = HttpStatus.FOUND)
-    public List<UserEntity> getFriends (@PathVariable String id){
-        List<FriendsEntity> listFriend = friendsRepository.findByUserId(id);
-        List<String> listUser = new ArrayList<>();
-        for (FriendsEntity f : listFriend){
-            listUser.add(f.getFriendId());
+    public List<UserEntity> getFriends (@PathVariable String id , @PathVariable String token){
+        TokensEntity tokensEntity = tokensRepository.findByToken(token);
+        if (tokensEntity != null && tokensEntity.getUserId().equals(id) ) {
+            List<FriendsEntity> listFriend = friendsRepository.findByUserId(id);
+            List<String> listUser = new ArrayList<>();
+            for (FriendsEntity f : listFriend) {
+                listUser.add(f.getFriendId());
+            }
+            return userRepository.findByUserNameIn(listUser);
+        }else{
+            return null;
         }
-        return userRepository.findByUserNameIn(listUser);
     }
 
-    @PostMapping(path = "/", consumes = "application/json", produces = "application/json")
+    @PostMapping(path = "/{token}", consumes = "application/json", produces = "application/json")
     @ResponseStatus(code = HttpStatus.CREATED)
-    public FriendsEntity addFriend (@RequestBody FriendsEntity entity){
-        return friendsRepository.save(entity);
+    public FriendsEntity addFriend (@RequestBody FriendsEntity entity , @PathVariable String token){
+        TokensEntity tokensEntity = tokensRepository.findByToken(token);
+        if (tokensEntity != null && tokensEntity.getUserId().equals(entity.getUserId()) ) {
+            return friendsRepository.save(entity);
+        }else{
+            return null;
+        }
     }
 
-    @PostMapping (path = "/delete" , consumes = "application/json")
+    @PostMapping (path = "/delete/{token}" , consumes = "application/json")
     @ResponseStatus(code = HttpStatus.OK)
-    public void deletePost(@RequestBody FriendsId entity) {
-        friendsRepository.deleteById(entity);
+    public void deleteFriend(@RequestBody FriendsId entity , @PathVariable String token) {
+        TokensEntity tokensEntity = tokensRepository.findByToken(token);
+        if (tokensEntity != null && tokensEntity.getUserId().equals(entity.getUserId()) ) {
+            friendsRepository.deleteById(entity);
+        }
     }
 
     @GetMapping(path = "/isExist/{userId}/{friendId}" )
