+++ src/main/java/com/socialhk/social_network/controller/PostController.java
@@ -1,11 +1,9 @@
 package com.socialhk.social_network.controller;
 
-import com.socialhk.social_network.model.entity.FriendsEntity;
-import com.socialhk.social_network.model.entity.PostEntity;
-import com.socialhk.social_network.model.entity.PostId;
-import com.socialhk.social_network.model.entity.UserEntity;
+import com.socialhk.social_network.model.entity.*;
 import com.socialhk.social_network.model.repository.FriendRepository;
 import com.socialhk.social_network.model.repository.PostRepository;
+import com.socialhk.social_network.model.repository.TokensRepository;
 import com.socialhk.social_network.model.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
@@ -27,18 +25,25 @@ public class PostController {
     private FriendRepository friendsRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private TokensRepository tokensRepository;
 
 
-    @GetMapping(path ={"/getAll" ,"/getAll/{user}" , "/getAll/{user}/{page}"}, produces = "application/json")
+    @GetMapping(path ={"/getAll" ,"/getAll/{user}/{token}" , "/getAll/{user}/{page}"}, produces = "application/json")
     @ResponseStatus(code = HttpStatus.FOUND)
-    public List<PostEntity> getAllPost(@PathVariable(required = false) Integer page , @PathVariable String user ) {
-        page = (page == null)? 1 : page;
-        int last = page * 10;
-        Pageable pageable = PageRequest.of(last - 10 , last, Sort.Direction.DESC,"date");
+    public List<PostEntity> getAllPost(@PathVariable(required = false) Integer page , @PathVariable String user,@PathVariable String token ) {
+        TokensEntity tokensEntity = tokensRepository.findByToken(token);
+        if (tokensEntity != null && tokensEntity.getUserId().equals(user) ) {
+            page = (page == null) ? 1 : page;
+            int last = page * 10;
+            Pageable pageable = PageRequest.of(last - 10, last, Sort.Direction.DESC, "date");
 
-        List<String> friends = getFriends(user);
-        System.out.println(friends);
-        return repository.findByOwnerIdIn(friends,pageable);
+            List<String> friends = getFriends(user);
+            System.out.println(friends);
+            return repository.findByOwnerIdIn(friends, pageable);
+        }else{
+            return null;
+        }
     }
 
 
@@ -83,25 +88,38 @@ public class PostController {
 
     }
 
-    @PostMapping(path = "/", consumes = "application/json", produces = "application/json")
+    @PostMapping(path = "/{token}", consumes = "application/json", produces = "application/json")
     @ResponseStatus(code = HttpStatus.CREATED)
-    public PostEntity addPost(@RequestBody PostEntity post) {
-        return repository.save(post);
-    }
-
-    @PostMapping(path = "/edit/{oldTitle}", consumes = "application/json", produces = "application/json")
-    @ResponseStatus(code = HttpStatus.CREATED)
-    public PostEntity editPost(@RequestBody PostEntity post, @PathVariable String oldTitle) {
-        if(!post.getTitle().equals(oldTitle)){
-            repository.deleteById( new PostId(post.getOwnerId(),oldTitle) );
+    public PostEntity addPost(@RequestBody PostEntity post , @PathVariable String token) {
+        TokensEntity tokensEntity = tokensRepository.findByToken(token);
+        if (tokensEntity != null && tokensEntity.getUserId().equals(post.getOwnerId()) ) {
+            return repository.save(post);
+        }else{
+            return null;
         }
-        return repository.save(post);
     }
 
-    @DeleteMapping (path = "/delete/{id}/{title}")
+    @PostMapping(path = "/edit/{oldTitle}/{token}", consumes = "application/json", produces = "application/json")
+    @ResponseStatus(code = HttpStatus.CREATED)
+    public PostEntity editPost(@RequestBody PostEntity post, @PathVariable String oldTitle , @PathVariable String token) {
+        TokensEntity tokensEntity = tokensRepository.findByToken(token);
+        if (tokensEntity != null && tokensEntity.getUserId().equals(post.getOwnerId()) ) {
+            if (!post.getTitle().equals(oldTitle)) {
+                repository.deleteById(new PostId(post.getOwnerId(), oldTitle));
+            }
+            return repository.save(post);
+        }else{
+            return null;
+        }
+    }
+
+    @DeleteMapping (path = "/delete/{id}/{title}/{token}")
     @ResponseStatus(code = HttpStatus.OK)
-    public void deletePost(@PathVariable String id,@PathVariable String title) {
-         repository.deleteById(new PostId(id,title));
+    public void deletePost(@PathVariable String id,@PathVariable String title , @PathVariable String token) {
+        TokensEntity tokensEntity = tokensRepository.findByToken(token);
+        if (tokensEntity != null && tokensEntity.getUserId().equals(id) ) {
+            repository.deleteById(new PostId(id, title));
+        }
     }
 
     @GetMapping(path = "/isExist/{id}/{title}" )
